requirements.txt
torch
torchvision
numpy
opencv-python
Pillow

matplotlib

altair==4.2.2
streamlit